backend/requirements.txt
# Core FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# AI/ML Dependencies
torch==2.10.0
torchaudio==2.10.0
openai-whisper
soundfile==0.12.1
librosa==0.10.2.post1
numpy==1.26.4
pyttsx3==2.98
pyspellchecker==0.8.1

# Data validation and serialization
pydantic>=2.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1

# Environment and configuration
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2
